requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-core==2.14.1
kafka-python==2.0.2
aiokafka==0.10.0
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
structlog==24.1.0